+++ tests/Meta_test.ml
@@ -21,7 +21,8 @@ let t = T.create
 *)
 let shell_command ?(expected_exit_code = 0) ~__LOC__:loc command =
   printf "RUN %s\n%!" command;
-  let exit_code = Sys.command command in
+  let bash_command = sprintf "bash -c '%s'" command in
+  let exit_code = Sys.command bash_command in
   if exit_code <> expected_exit_code then
     failwith
       (sprintf "%s:\nCommand '%s' exited with code %i but code %i was expected."
@@ -112,6 +113,8 @@ let clear_snapshots ~__LOC__:loc () =
   shell_command ~__LOC__:loc "mkdir -p tests/custom-snapshots";
   shell_command ~__LOC__:loc "rm -f tests/custom-snapshots/*"
 
+(* FIXME: This integration test fails on Windows because the expected output
+   includes paths generated on POSIX systems. It runs correctly however. *)
 let test_standard_flow () =
   section "Clean start";
   clear_status ~__LOC__ ();
@@ -241,13 +244,25 @@ let remove_optional_lines =
 
 let mask_and_sort = mask_alcotest_output @ [ sort_lines; remove_optional_lines ]
 
+(* FIXME: Running parallel jobs on Windows is unstable, producing
+   Sys_error("Invalid argument")` during some attempts to flush channels.
+   Perhaps related to https://github.com/ocaml/ocaml/issues/13586 *)
+let skipped =
+  if Sys.win32 then
+    Some "Running tests in parallel is unstable on Windows"
+  else
+    None
+
 let tests =
   [
-    t ~checked_output:(T.stdxxx ()) ~normalize:mask_alcotest_output
+    t ~checked_output:(T.stdxxx ())
+      ~normalize:mask_alcotest_output
       "standard flow" test_standard_flow;
     t ~checked_output:(T.stdxxx ()) ~normalize:mask_and_sort
+      ?skipped
       "fewer workers than tests" test_fewer_workers_than_tests;
     t ~checked_output:(T.stdxxx ()) ~normalize:mask_and_sort
+      ?skipped
       "more workers than tests" test_more_workers_than_tests;
     t "failing flow run"
       ~expected_outcome:
@@ -275,6 +290,12 @@ let tests =
   ]
 
 let () =
+  (* stdout and stderr are in "text mode" by default, and on windows this
+     entails rewriting line endings to CLRF. This makes the test output
+     incompatible between Windows and POSIX platforms. Setting the channels to
+     binary mode ensures consistent output. *)
+  set_binary_mode_out stdout true;
+  set_binary_mode_out stderr true;
   (* We have a few tests that use the same workspace. To avoid conflicts,
      we run them sequentially. *)
   Testo.interpret_argv ~default_workers:(Some 0)